requirements.txt
pyserial>=3.5
xmodem>=0.4.7
colorama>=0.4.4
